Ajout d'un TU pour l'ajout d'un nouvel utilisateur ainsi que l'implémentation de la méthode, test ok

+++ users-ms/src/main/java/com/medhead/users_ms/entities/User.java
@@ -9,7 +9,7 @@ public class User {
 
     @Id
     @GeneratedValue
-    private Integer userId ;
+    private Long userId ;
     private String pseudo ;
     private String password ;
     private String email ;
@@ -18,7 +18,7 @@ public class User {
     public User() {
     }
 
-    public User(Integer userId, String pseudo, String password, String email, Boolean activated) {
+    public User(Long userId, String pseudo, String password, String email, Boolean activated) {
         this.userId = userId;
         this.pseudo = pseudo;
         this.password = password;
@@ -26,11 +26,11 @@ public class User {
         this.activated = activated;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
